+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+### v4.1.0 (2022-07-23)
+
+- feat: refactoring component.
+- feat: export `Props` `Events` `DEFAULT_CONFIG`.
+- chore: upgrade deps and upgrade vite to v3.
+
 ### v4.0.2 (2022-05-11)
 
 - Fix CSS `z-index` for `:not(:last-child)`